pyproject.toml
[tool.poetry]
name = "flipcase-web"
version = "0.1.0"
description = ""
authors = ["Javier Javier <javierjdaza@gmail.com>"]
readme = "README.md"
packages = [{include = "flipcase_web"}]

[tool.poetry.dependencies]
python = "^3.10"
streamlit = "^1.28.1"
Pillow = "^10.1.0"
streamlit-option-menu = "^0.3.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
